Extract nested dict and param tree
